🚚 Move file utilities function to /documents
