+++ src/documents/index.js
@@ -1,0 +1,1 @@
+export {default as DocumentRoutes} from './routes';
